terraform/gcp/functions/auto-failover/requirements.txt
google-cloud-compute==1.14.1
google-cloud-logging==3.5.0
google-cloud-monitoring==2.15.1
requests==2.31.0